health_assistant/pyproject.toml
[tool.poetry]
name = "health-assistant-ui"
version = "0.1.0"
description = ""
authors = ["Goutham523 <gouthamktm005@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.52.1"
requests = "^2.32.5"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
